poll-backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==3.8.0
celery==5.3.6
django-redis==5.4.0
